code/ldopa/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
asn1crypto=0.22.0=py27h94ebe91_1
backports=1.0=py27h63c9359_1
backports.csv=1.0.1=py27_0
backports.weakref=1.0rc1=py27_0
bleach=1.5.0=py27_0
ca-certificates=2017.08.26=h1d4fec5_0
certifi=2017.7.27.1=py27h9ceb091_0
cffi=1.10.0=py27hf1aaaf4_1
chardet=3.0.4=py27hfa10054_1
cryptography=2.0.3=py27hea39389_1
cudatoolkit=8.0=3
cudnn=6.0.21=cuda8.0_0
enum34=1.1.6=py27h99a27e9_1
funcsigs=1.0.2=py27h83f16ab_0
future=0.16.0=py27_1
h5py=2.7.0=np112py27_0
hdf5=1.8.17=2
html5lib=0.9999999=py27_0
idna=2.6=py27h5722d68_1
intel-openmp=2018.0.0=h15fc484_7
ipaddress=1.0.18=py27h337fd85_0
jbig=2.1=hdba287a_0
joblib=0.11=py27_0
jpeg=8d=2
keras=2.0.6=py27_0
libedit=3.1=heed3624_0
libffi=3.2.1=h4deb6c0_3
libgcc=7.2.0=h69d50b8_2
libgcc-ng=7.2.0=h7cc24e2_2
libgfortran=3.0.0=1
libgfortran-ng=7.2.0=h9f7466a_2
libgpuarray=0.7.4=0
libpng=1.6.32=hda9c8bc_2
libprotobuf=3.4.0=0
libstdcxx-ng=7.2.0=h7a57d05_2
libtiff=4.0.6=2
mako=1.0.7=py27_0
markdown=2.6.9=py27_0
markupsafe=1.0=py27_0
mkl=2017.0.4=h4c4d0af_0
mock=2.0.0=py27h0c0c831_0
ncurses=6.0=h06874d7_1
numpy=1.12.1=py27_0
opencv=3.1.0=np112py27_1
openssl=1.0.2l=h077ae2c_5
pandas=0.20.3=py27hfd1eabf_2
pbr=3.1.1=py27hf64632f_0
pip=9.0.1=py27ha730c48_4
progressbar=2.3=py27_0
protobuf=3.4.0=py27_0
pycparser=2.18=py27hefa08c5_1
pygpu=0.7.4=py27_0
pyopenssl=17.2.0=py27h189ff3b_0
pysocks=1.6.7=py27he2db6d2_1
python=2.7.14=h89e7a4a_22
python-dateutil=2.6.1=py27h4ca5741_1
pytz=2017.2=py27hcac29fa_1
pyyaml=3.12=py27_1
readline=7.0=hac23ff0_3
requests=2.18.4=py27hc5b0589_1
scikit-learn=0.19.0=np112py27_0
scipy=0.19.1=np112py27_0
setuptools=36.5.0=py27h68b189e_0
six=1.11.0=py27h5f960f1_1
sqlite=3.20.1=h6d8b0f3_1
synapseclient=1.7.1=py27_0
tensorflow-gpu=1.3.0=0
tensorflow-gpu-base=1.3.0=py27cuda8.0cudnn6.0_1
tensorflow-tensorboard=0.1.5=py27_0
theano=0.9.0=py27_1
tk=8.6.7=h5979e9b_1
urllib3=1.22=py27ha55213b_0
werkzeug=0.12.2=py27hbf75dff_0
wheel=0.29.0=py27h411dd7b_1
xz=5.2.3=h2bcbf08_1
yaml=0.1.6=0
zlib=1.2.11=hfbfcf68_1
